backend/pyproject.toml
[project]
name = "auction-ocr-backend"
version = "0.1.0"
description = "Auction OCR API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.3",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "celery>=5.4",
  "redis>=5.0",
  "boto3>=1.34",
  "python-multipart>=0.0.9",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "email-validator>=2.2",
  "python-dateutil>=2.9",
  "pillow>=10.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
